Parse

Description

Parser combinators for TransmormeR, using the Text.Parsec library. This is work in progress, and the final goal is to implement a subset of R grammar with (refinement) type annotations. Refer to the paper/design doc on Github for the definition of the grammar.

Documentation

parse :: Parser a -> SourceName -> String -> Either ParseError a

run a parser on an input string. parse value "t.R" "1.23" will return (hopefully!) Right (NUM 1.23)

prog :: Parser Transformation

parse a TransformeR program

expr :: Parser Expression

parse an expression, which can be a term or some operation over terms.

term :: Parser Expression

parse a term

value :: Parser Value

parse a value

opr :: Parser OPR

parse an operator, consider to integrate into table

table :: OperatorTable String u Identity Expression

table of operators with fixities and precedence. note that field access is treated as a post-fix operator

recsig :: Parser Sig

parse a record-type signature

tysig :: Parser Descriptor

parse a primitive type signature