Syntax

Description

Grammar for TransmormeR, with abstract and concrete syntax (the latter defined in Show instances). This is work in progress, and the final goal is to implement a subset of R grammar with (refinement) type annotations. Refer to the paper/design doc on Github for the definition of the grammar.

Documentation

data Transformation

A transformation has a name, is signed by an input- and a return-data descriptor and has a body.

Constructors

TRANS Name Sig Sig Expression

data Sig

Record-type encodes column names and their descriptors

Constructors

Sig [(Name, Descriptor)]

data Expression

expressions that can appear in the transformation body: a literal value, a field access, an operator application, a record, a variable, a field mutation, sequencing of expressions, or a variable assignment. The order might matter during pretty printing (operator precedence)!

Constructors

LIT Value
literal value

PROJ Expression Field
field access

OP OPR [Expression]
operator application

REC [(Field, Expression)]
record

VAR Name
variable

MUTATE Name Field Expression
field mutation

SEQ Expression Expression
sequencing of exps

ASSIGN Name Expression
variable assignment

type Descriptor = Range

a descriptor encodes range

data Range

range encodes information neccessary to infer the sensitivity (together with the DP-mechanism)

Constructors

INTERVAL Number Number
currently using closed intervals for floats

SET [Name]
a set of categories (bins in a histogram)

data Value

a value is either

Constructors

NUM Number
a number (currently floats only)

CAT Category
a category (bin name in a histogram)

MAP [(Field, Value)]
a record

data OPR

supported statistical operators

Constructors

SUM
sum 2 numbers

CONCAT
concatinate 2 categories