Types

Description

Type System for TransmormeR, This is work in progress, and the final goal is to implement a type system consistent with R semantics. There are 2 ways to do that: implement a interpreter from the semantics defined in the paper/design doc, and loop {test, change} until you massage the semantics so that the interpreter behaves the same on all inputs you can think of just like R does. Or, you can generate a bunch of well-typed TransformeR programs and run them under R, modify TransformeR until all outputs have the type that TransformeR claims them to have. Refer to the paper/design doc on Github for the definition of the grammar.

Documentation

class RefTy t where

types are refined with ranges. refinement types should implement a check for well-typedness

Methods

tyOK :: t -> Bool

data RecTy

a record type

Constructors

RecTy [(Label, Ty)]

data Ty

a primitive type

Constructors

NumTy (Number, Number)

CatTy [Category]

type Gamma = [(Name, Either RecTy Ty)]

type environment stores both record type and primitive types. A better design would replace Either with a home-made datatype, like data Ty = PT ExTy | RT RecTy and have Gamma just store Ty. Gamma also deserves a better data structure, perhaps Data.Map?

mapsTo :: [(t, t1)] -> t -> t1 -> [(t, t1)]

update the type environment

data TransTy

valid transformations have a type

Constructors

TransTy RecTy RecTy

tyTrans :: Transformation -> TransTy

type-check a transformation

ty :: Gamma -> Expression -> (Either RecTy Ty, Gamma)

type-check an expression

tyJoin :: OPR -> [Either t Ty] -> Ty

operator typing rules. Serves as a basis for all statistical operations.